added set scores, but not complete
